requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
idna==3.4
novas==3.1.1.5
Pillow==9.4.0
pycodestyle==2.10.0
pytz==2023.2
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
